Bug 1069646: scale frame rate initialization in webrtc media_opimization r=gcp
UltraBlame original commit: 148862fe121eba7856927c27e88af07b116babcf

+++ media/webrtc/trunk/webrtc/modules/video_coding/main/source/content_metrics_processing.h
@@ -45,7 +45,7 @@ class VCMContentMetricsProcessing {
   int Reset();
 
   
-  void UpdateFrameRate(uint32_t frameRate);
+  void UpdateFrameRate(float frameRate);
